Add option to press enter to exit
This adds an option to the configuration file to ask for confirmation
before exiting. It defaults to false. If True it will require the user
to press the enter key before the program will exit, allowing the user
to view the summary results on screen before the window closes if run
outside the command prompt.